Add functionality to all buttons but equals.
All buttons now cause interactions with strings which represent, and are displayed
as required on the upper and lower sections of the screen. All numbers and operators
are also stored in an array for evaluation when equals is pressed (not yet
implamented). Contradictory interactions, such as pressing two operators in a row,
or trying to have multiple decimal points in one number have been filtered out.