backspaceFunc added, called by backspcae button.
backspaceFunc initially clears the workingString if there is anything in it,
and resets the display to 0. If there is nothing in workingString (either hasn't
added to or has already been clearer) the last operator and number are removed
from the rawMath array, and the screenUpper display changed to reflect this.