Add equals, resolve, anf functions for each op.
Individual functions solely do the work of each operation. These are called
as required by the resolve function depending on what operator it has
encountered as it moves through the rawMath array. The resolve function
works recursively.
The resolve function itself is repeatedly called by the equals function
which tells resolve what to look for, applying the order of operations
to the calls. The equals function also deals with the strings and output
text.