Add divide by 0 check. Remove leading 0s.
divZeroCheck was added to ensure that no attempt will be made to divide by zero.
It is called whenever an op button is pressed, so that /0 cannot even be
entered into rawMath. It is also called by equals, so that the last digit
entered cannot be a 0 after a /.
numButtonFunc was modified to remove leading 0s from any numbers on the screen.